arquitectura y conexion completas

+++ app/src/main/java/com/njm/carpintime/domain/model/Temp.kt
@@ -4,7 +4,7 @@ data class Temp(
     var day: Double,
     var eve: Double,
     var max: Double,
-    var min: Int,
+    var min: Double,
     var morn: Double,
-    var night: Int
+    var night: Double
 )
